Update to number of pings to initialize net
While testing this in Haiti on virtual machines, we found that pinging 4 times was not enough (4/6 machines failed to initialize the net in time).  The machines that didn't get it up in time failed and they had to restart.  We upped it to 30 or 40.  We found that it's not alot of time to add to the script to ensure it fires 100%